+++ src/main/java/com/example/tiggle/service/account/AccountService.java
@@ -12,11 +12,11 @@ public interface AccountService {
     
     Mono<OneWonVerificationResponse> sendOneWonVerification(String encryptedUserKey, String accountNo);
     
-    Mono<OneWonVerificationValidateResponse> validateOneWonAuth(String encryptedUserKey, String accountNo, String authCode, Integer userId);
+    Mono<OneWonVerificationValidateResponse> validateOneWonAuth(String encryptedUserKey, String accountNo, String authCode, Long userId);
     
-    Mono<ApiResponse<Void>> registerPrimaryAccount(String accountNo, String verificationToken, Integer userId);
+    Mono<ApiResponse<Void>> registerPrimaryAccount(String accountNo, String verificationToken, Long userId);
     
-    Mono<ApiResponse<PrimaryAccountInfoDto>> getPrimaryAccount(String encryptedUserKey, Integer userId);
+    Mono<ApiResponse<PrimaryAccountInfoDto>> getPrimaryAccount(String encryptedUserKey, Long userId);
     
     Mono<ApiResponse<AccountHolderInfoDto>> getAccountHolder(String encryptedUserKey, String accountNo);
